[RENAME] (common) AuthUser / CustomException 패키지 이동
