+++ src/main/java/com/example/lastproject/domain/auth/service/AuthService.java
@@ -4,7 +4,7 @@ import com.example.lastproject.domain.auth.dto.request.SigninRequest;
 import com.example.lastproject.domain.auth.dto.request.SignupRequest;
 import com.example.lastproject.domain.auth.dto.response.SignupResponse;
 import com.example.lastproject.domain.auth.dto.response.WithdrawalResponse;
-import com.example.lastproject.domain.auth.entity.AuthUser;
+import com.example.lastproject.common.dto.AuthUser;
 
 public interface AuthService {
