+++ src/main/java/com/example/lastproject/domain/penalty/controller/PenaltyController.java
@@ -1,6 +1,6 @@
 package com.example.lastproject.domain.penalty.controller;
 
-import com.example.lastproject.domain.auth.entity.AuthUser;
+import com.example.lastproject.common.dto.AuthUser;
 import com.example.lastproject.domain.penalty.dto.request.PenaltyRequest;
 import com.example.lastproject.domain.penalty.dto.response.PenaltyResponse;
 import com.example.lastproject.domain.penalty.service.PenaltyServiceImpl;
