+++ src/main/java/com/example/lastproject/domain/penalty/service/PenaltyService.java
@@ -1,6 +1,6 @@
 package com.example.lastproject.domain.penalty.service;
 
-import com.example.lastproject.domain.auth.entity.AuthUser;
+import com.example.lastproject.common.dto.AuthUser;
 import com.example.lastproject.domain.penalty.dto.request.PenaltyRequest;
 
 public interface PenaltyService {
