+++ src/main/java/com/example/lastproject/domain/penalty/service/PenaltyServiceImpl.java
@@ -1,8 +1,8 @@
 package com.example.lastproject.domain.penalty.service;
 
-import com.example.lastproject.common.CustomException;
+import com.example.lastproject.common.exception.CustomException;
 import com.example.lastproject.common.enums.ErrorCode;
-import com.example.lastproject.domain.auth.entity.AuthUser;
+import com.example.lastproject.common.dto.AuthUser;
 import com.example.lastproject.domain.party.entity.Party;
 import com.example.lastproject.domain.party.enums.PartyStatus;
 import com.example.lastproject.domain.party.repository.PartyRepository;
